--- a/rpcs3/Emu/RSX/GL/GLGSRender.cpp
+++ b/rpcs3/Emu/RSX/GL/GLGSRender.cpp
@@ -781,7 +781,10 @@ bool GLGSRender::load_program()
 	if (shadermode != shader_mode::interpreter_only) [[likely]]
 	{
 		void* pipeline_properties = nullptr;
-		std::tie(m_program, m_vertex_prog, m_fragment_prog) = m_prog_buffer.get_graphics_pipeline(current_vertex_program, current_fragment_program, pipeline_properties,
+		std::tie(m_program, m_vertex_prog, m_fragment_prog) = m_prog_buffer.get_graphics_pipeline(
+			current_vertex_program, vertex_program_invalidation_count, 
+			current_fragment_program, fragment_program_invalidation_count,
+			pipeline_properties,
 			shadermode != shader_mode::recompiler, true);
 
 		if (m_prog_buffer.check_cache_missed())