--- a/rpcs3/Emu/RSX/NV47/HW/nv406e.cpp
+++ b/rpcs3/Emu/RSX/NV47/HW/nv406e.cpp
@@ -25,6 +25,9 @@ namespace rsx
 			RSX(ctx)->sync_point_request.release(true);
 			const u32 addr = get_address(REGS(ctx)->semaphore_offset_406e(), REGS(ctx)->semaphore_context_dma_406e());
 
+			// Syncronization point, may be associated with memory changes without actually changing addresses
+			RSX(ctx)->m_graphics_state |= rsx::pipeline_state::fragment_program_ucode_dirty;
+
 			const auto& sema = vm::_ref<RsxSemaphore>(addr).val;
 
 			if (sema == arg)