--- a/rpcs3/Emu/RSX/NV47/HW/nv47_sync.hpp
+++ b/rpcs3/Emu/RSX/NV47/HW/nv47_sync.hpp
@@ -44,6 +44,9 @@ namespace rsx
 						// Manually flush the pipeline.
 						// It is possible to stream report writes using the host GPU, but that generates too much submit traffic.
 						RSX(ctx)->sync();
+
+						// Syncronization point, may be associated with memory changes without actually changing addresses
+						RSX(ctx)->m_graphics_state |= rsx::pipeline_state::fragment_program_ucode_dirty;
 					}
 				}