--- a/rpcs3/Emu/RSX/RSXThread.cpp
+++ b/rpcs3/Emu/RSX/RSXThread.cpp
@@ -1904,6 +1904,7 @@ namespace rsx
 		}
 
 		m_graphics_state.clear(rsx::pipeline_state::fragment_program_ucode_dirty);
+		fragment_program_invalidation_count++;
 
 		// Request for update of fragment constants if the program block is invalidated
 		m_graphics_state |= rsx::pipeline_state::fragment_constants_dirty;
@@ -2046,6 +2047,7 @@ namespace rsx
 		ensure(!m_graphics_state.test(rsx::pipeline_state::fragment_program_ucode_dirty));
 
 		m_graphics_state.clear(rsx::pipeline_state::fragment_program_dirty);
+		fragment_program_invalidation_count++;
 
 		current_fragment_program.ctrl = m_ctx->register_state->shader_control() & (CELL_GCM_SHADER_CONTROL_32_BITS_EXPORTS | CELL_GCM_SHADER_CONTROL_DEPTH_EXPORT);
 		current_fragment_program.texcoord_control_mask = m_ctx->register_state->texcoord_control_mask();