--- a/rpcs3/Emu/RSX/RSXThread.cpp
+++ b/rpcs3/Emu/RSX/RSXThread.cpp
@@ -1953,6 +1953,7 @@ namespace rsx
 		}
 
 		m_graphics_state.clear(rsx::pipeline_state::vertex_program_ucode_dirty);
+		vertex_program_invalidation_count++;
 
 		// Reload transform constants unconditionally for now
 		m_graphics_state |= rsx::pipeline_state::transform_constants_dirty;
@@ -2010,6 +2011,7 @@ namespace rsx
 			return;
 		}
 
+		vertex_program_invalidation_count++;
 		ensure(!m_graphics_state.test(rsx::pipeline_state::vertex_program_ucode_dirty));
 		current_vertex_program.output_mask = rsx::method_registers.vertex_attrib_output_mask();