--- a/rpcs3/Emu/RSX/VK/VKProgramBuffer.h
+++ b/rpcs3/Emu/RSX/VK/VKProgramBuffer.h
@@ -88,13 +88,13 @@ namespace vk
 		template <typename... Args>
 		void add_pipeline_entry(RSXVertexProgram& vp, RSXFragmentProgram& fp, vk::pipeline_props& props, Args&& ...args)
 		{
-			get_graphics_pipeline(vp, fp, props, false, false, std::forward<Args>(args)...);
+			get_graphics_pipeline(vp, umax, fp, umax, props, false, false, std::forward<Args>(args)...);
 		}
 
 		void preload_programs(const RSXVertexProgram& vp, const RSXFragmentProgram& fp)
 		{
-			search_vertex_program(vp);
-			search_fragment_program(fp);
+			search_vertex_program(vp, umax);
+			search_fragment_program(fp, umax);
 		}
 
 		bool check_cache_missed() const