… rpcs3/Emu/RSX/rsx_methods.cpp