* Prevent reload of vertex transform program if program data write did not change the program.
* Optimize sucessive fragment program and vertex program use in two ways:
1. If no change was made to the shader program, try to reuse the compiled program output without heavy validation.
2. If data partially matches between current vertex and shader programs to the previous ones, try to do a fast comparison to see f they match. Skipping hash table resolving.
* Add a method that encapsulates fragment texture changes which is a minor optimization of FIFO execution.

**Metal Gear Online (Menu): ~7000 FPS to 10,000 fps**
**Ninja Guiden Sigma:  194 fps to 202 fps**
**Ratchet And Clank: Toold Of Destruction: 22 -> 24 FPS!**