@elad335 2025-03-03
I have removed some other state-management from storage hashing.