@cipherxof 2025-03-03
After https://github.com/RPCS3/rpcs3/pull/16797/commits/55c6f021ea9afc371608e99f4bee6449b51dd8d3 (was about 7,000fps before this commit):

![Screenshot From 2025-03-02 22-56-03](https://github.com/user-attachments/assets/723d25bc-19ec-4b1b-bb81-19a177a14615)

![image](https://github.com/user-attachments/assets/98b404a1-8669-417a-8512-a5b4dbbcedc2)

Seems about the same everywhere else that I've tested compared to https://github.com/RPCS3/rpcs3/pull/16780