@Whatcookie 2025-03-03
Latest build is still compiling shaders on every run
![image](https://github.com/user-attachments/assets/c6fa6e5e-c6a4-4db1-bc1a-0e9c3bdf450e)