@Whatcookie 2025-03-04
Provides a nice uplift when avx-512-icl is enabled, and an even bigger uplift for those without.

With AVX-512-ICL (~204->209avg)
![image](https://github.com/user-attachments/assets/ae4fe49c-07f6-4151-8472-068e94025dcd)

With AVX-512-ICL paths disabled at compile time (~194->202avg)
![image](https://github.com/user-attachments/assets/35151b5a-ae6b-41c5-b1f0-5f73973e8b2a)