@Whatcookie 2025-03-04
uplift is very good in ratchet tod
![image](https://github.com/user-attachments/assets/ea62cbea-13b4-46e4-800c-44a0aee21a1f)