@cipherxof 2025-03-04
> I decided to go on a different route at the end - if the last program cached seems very similar to the current one, do not do hashing against the hash table, instead compare the two programs directly and return them if they match precisely. The drawback is of course, that fast likely-ness test is not catching all cases of mismatch and thus this may add a second comparison in edge cases. May benefit RaC ToD which does not invalidate FP and VP programs down its pipeline.
> 
> I also added a minor optimization for VP UCODE invalidation checks.
> 
> @cipherxof can you test it?

Performance is the same as my last test :+1: