@elad335 2025-03-05
I have pushed another optimization, this time if VP shader has not been changed since last VP shader fetch we skip the comparison altogether. This is trivial for VP as its changes are placed on the instruction stream.

May benefit RaC ToD and other games.