@elad335 2025-03-06
Pushed another optimization that reduces FP shader comparison between sequencing draws.